chore(db): update column types

+++ packages/database/schemas/stripped/manager.sql
@@ -1,7 +1,7 @@
 USE `manager`;
 
 CREATE TABLE `botShardStat` (
-  `shardId` mediumint NOT NULL,
+  `shardId` int NOT NULL,
   `status` tinyint NOT NULL,
   `serverCount` int NOT NULL DEFAULT '0',
   `uptimeSeconds` int NOT NULL DEFAULT '0',
@@ -23,7 +23,7 @@ CREATE TABLE `botShardStat` (
 );
 
 CREATE TABLE `dbShard` (
-  `id` bigint NOT NULL DEFAULT '0',
+  `id` smallint NOT NULL DEFAULT '0',
   `host` varchar(45) NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`)
 );
